+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## 1.1.2
-- Increase the number of clones to 120
+- Increase the number of clones to 130
 
 ## 1.1.1
 - Generate random parameters for `u`, `mean` and `std` if not provided.